fix: add opacity 0 (#168)
Slide and Zoom flicker when the element is just entering the view and
certain conditions are met. Slide flickers when direction is up.
Zoom flickers when it is styled. This bug is shown in this Code Sandbox,
https://codesandbox.io/s/react-awesome-reveal-bug-vy35p2 .

This is fixed by applying opacity: 0 when the element is not in view.

resolve #168